+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 - [\#337](https://github.com/cosmos/evm/pull/337) Support state overrides in eth_call.
 - [\#502](https://github.com/cosmos/evm/pull/502) Add block time in derived logs.
 - [\#633](https://github.com/cosmos/evm/pull/633) go-ethereum metrics are now emitted on a separate server. default address: 127.0.0.1:8100.
+- [\#650](https://github.com/cosmos/evm/pull/650) Make staking precompile queries return the full validators' description structure.
 
 ### STATE BREAKING